btrfs: fix memory management in find_and_setup_log_root()
log_root does not have to be freed at the end because it is assigned to
fs_info->log_root_tree.
However, it have to be freed when the function returns on an error. Idem
for log_root->node.

I have based my modifications on the most recent btrfs's commit, at
http://git.kernel.org/cgit/linux/kernel/git/mason/btrfs-progs.git/tree/disk-io.c?id=650e656a8b9c1fbe4ec5cd8c48ae285b8abd3b69

I have ordered the lines differently to make the code a little bit simpler.

Signed-off-by: Patryck Rouleau <pfrouleau@gmail.com>