Fix validation of image without checksums
When the image does not use a checksum, it is validated with custom code.
Some file systems do not report the same used block count than what partclone
find while he build the used bitmap. We must use the count from the bitmap
because the clone and restore are based solely on the bitmap.

Signed-off-by: Patrick Rouleau <pfrouleau@gmail.com>