Support compressed cloned images
The ncurses mode obviously conflicts with writing the cloned image
to stdout. Piping the image to a command to compress it while in
ncurses mode is unusable because of this.

Fix it by adding a new option -x / --compresscmd CMD that is
executed in a write-only pipe internally by partclone.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>