fail-mbr: Remove binutils section .note.gnu.build-id using objcopy
GNU Binutils 2.36 configures with --enable-x86-used-note by default for
Linux/x86. This adds the new section .note.gnu.build-id, which needs to
be removed.

When building fail-mbr.o with current binutils, ld links the file into
an executable format. However this is not accepted by current binutils
for building fail-mbr.image; thus it requires -Wl,--relocatable for the
former and -Wl,--no-dynamic-linker for the latter.

See also:
- https://sourceware.org/pipermail/binutils/2021-January/115071.html
- https://sourceware.org/bugzilla/show_bug.cgi?id=26223
- https://sourceware.org/bugzilla/show_bug.cgi?id=27753
- https://lists.denx.de/pipermail/u-boot/2021-June/452303.html